// src/apis/types.service.ts
export interface ICityLatLong {
  lat: string;
  lon: string;
  display_name: string;
  class?: string;
  type: string;
}